+++ apps/server/src/lib/eventEmitter.ts
@@ -3,6 +3,8 @@ import { EventEmitter } from "events";
 import { HFile, ProcessData } from "@howdypix/shared-types";
 
 export interface EventTypes {
+  newDirectory: { root: string; hfile: HFile };
+  unlinkDirectory: { root: string; hfile: HFile };
   newFile: { root: string; hfile: HFile };
   removeFile: { root: string; hfile: HFile };
   processFile: { root: string; hfile: HFile };