+++ apps/webapp/cypress/integration/album.filter.spec.ts
@@ -32,6 +32,7 @@ context("Album page full with filters", () => {
     cy.get("body").contains("Camera Make").click();
 
     cy.findByText("make 1").click();
+    cy.findByText("make 1").click();
     cy.findByText("make 2").click();
 
     //= ===============================================