+++ apps/webapp/pages/login.tsx
@@ -1,7 +1,6 @@
 import * as React from "react";
 import { NextPage } from "next";
 import { LoginBox } from "../src/module/login/LoginBox";
-import { useStore } from "../src/context/store";
 
 type InitialProps = { namespacesRequired: string[] };