+++ libs/utils/lib/utils.ts
@@ -1,6 +1,6 @@
 import { Channel, ConsumeMessage, Options, Replies } from "amqplib";
 import { join, parse } from "path";
-import { HFile, QueueName } from "@howdypix/shared-types";
+import { HFile, QueueName, SupportedMime } from "@howdypix/shared-types";
 import debug from "debug";
 import { hjoin, thumbnailPath } from "./path";
 import { isArray, isNil, isObject, omitBy, reduce } from "lodash";
@@ -44,7 +44,7 @@ export function generateThumbnailUrls(
   return [200, 600, 1280, 2048].map((size) => ({
     width: size,
     height: null,
-    url: `${baseUrl}/static/${hjoin({
+    url: `${baseUrl}/files/${hjoin({
       ...hfile,
       file: `${parse(hfile.file as string).name}x${size}.jpg`,
     })}`,
@@ -56,7 +56,7 @@ export function generateFileUrl(baseUrl: string, hfile: HFile): string {
     throw new Error("You need to pass a file path to generate a URL.");
   }
 
-  return `${baseUrl}/static/${hjoin(hfile)}`;
+  return `${baseUrl}/files/${hjoin(hfile)}`;
 }
 
 export function isHowdypixPath(path = ""): boolean {
@@ -161,4 +161,17 @@ export function sortJson(json: {} | unknown[]): typeof json {
 
 export function sortJsonStringify(json: {}): string {
   return stringify(sortJson(json));
+}
+
+export function parentDir(dir = ""): string {
+  return parse(dir).dir ?? "";
+}
+
+export function isSupportedMime(mime: keyof SupportedMime): boolean {
+  const supportedMime: SupportedMime = {
+    "image/jpeg": true,
+    "image/png": true,
+  };
+
+  return supportedMime[mime] ?? false;
 }